ui_service/requirements.txt
psycopg2-binary==2.9.9
loguru==0.7.2
fastapi
werkzeug
uvicorn
Jinja2